requirements.txt
contourpy==1.2.0
cycler==0.12.1
DateTime==5.3
fonttools==4.46.0
GDAL==3.8.0
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy @ file:///D:/bld/numpy_1700874663201/work/dist/numpy-1.26.2-cp39-cp39-win_amd64.whl#sha256=7f969c1153e33a193010a6f19d39efb0de3ac8d7013da0c994e8ecffeeebba1c
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
seaborn==0.13.0
six==1.16.0
tzdata==2023.3
zipp==3.17.0
zope.interface==6.1
